Update solution of auto correct problem to match problem description
The problem description asked for suggested words with distance <= 2,
and missed the fact that only the suggestions with distance <= 2 AND
minimum distance should be output.

Since the contest is over, instead of updating the description, the solution
and test cases are updated to match the description. Submissions are
being re-tested.